requirements.txt
langchain
langchain-openai
langchain-core
langchain-community
langsmith
langgraph
pydantic
yfinance
google-colab
ipykernel
jupyter
openai
python-dotenv
fastapi
uvicorn
